refactor restore and add restore of issues
